requirements.txt
ffmpeg-python==0.2.0
openai-whisper==20240930
python-dotenv==1.0.0
